requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
alembic==1.12.1